requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
blas=1.0=mkl
bzip2=1.0.8=h7f98852_4
ca-certificates=2020.12.5=ha878542_0
cairo=1.14.12=h8948797_3
certifi=2020.12.5=py37h89c1867_1
cffi=1.14.5=py37hc58025e_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py37_0
dbus=1.13.14=hb2f20db_0
expat=2.2.6=he6710b0_0
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.0=h9420a91_0
freeglut=3.2.1=h9c3ff4c_2
freetype=2.10.2=h5ab3b9f_0
gettext=0.19.8.1=h0b5b191_1005
glib=2.63.1=h3eb4bd4_1
gputil=1.4.0=pyh9f0ad1d_0
graphite2=1.3.13=h58526e2_1001
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hc401514_3
icu=58.2=he6710b0_3
intel-openmp=2020.1=217
jasper=2.0.14=h07fcdf6_1
joblib=0.15.1=py_0
jpeg=9b=h024ee3a_2
kiwisolver=1.2.0=py37hfd86e86_0
ld_impl_linux-64=2.33.1=h53a641e_7
libblas=3.8.0=21_mkl
libcblas=3.8.0=21_mkl
libedit=3.1.20191231=h7b6447c_0
libffi=3.3=he6710b0_1
libflac=1.3.3=h9c3ff4c_1
libgcc-ng=9.3.0=h2828fa1_18
libgfortran=3.0.0=1
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=he1b5a44_1001
libgomp=9.3.0=h2828fa1_18
liblapack=3.8.0=21_mkl
libllvm10=10.0.1=he513fc3_3
libogg=1.3.4=h7f98852_1
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7f98852_1
libpng=1.6.37=hbc83047_0
libprotobuf=3.13.0.1=h8b12597_0
libsndfile=1.0.31=h9c3ff4c_0
libstdcxx-ng=9.3.0=h6de172a_18
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libvorbis=1.3.7=h9c3ff4c_0
libvpx=1.7.0=h439df22_0
libxcb=1.13=h1bed415_1
libxml2=2.9.10=he19cac6_1
llvmlite=0.35.0=py37h9d7f4d0_1
lz4-c=1.9.2=he6710b0_0
matplotlib=3.3.2=py37h89c1867_1
matplotlib-base=3.3.2=py37hc9afd2a_1
mkl=2020.1=217
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.1.0=py37h23d657b_0
mkl_random=1.1.1=py37h0573a6f_0
ncurses=6.2=he6710b0_1
ninja=1.9.0=py37hfd86e86_0
numba=0.52.0=py37hdc94413_0
numpy=1.20.1=py37haa41c4c_0
numpy-base=1.18.5=py37hde5b4d6_0
olefile=0.46=py37_0
onnx=1.8.0=py37ha8e9a8c_0
opencv=3.4.2=py37h6fd60c2_1
openssl=1.1.1j=h7f98852_0
pandas=1.1.5=py37hdc94413_0
patsy=0.5.1=py_0
pcre=8.44=he6710b0_0
pillow=7.1.2=py37hb39fc2d_0
pip=20.1.1=py37_1
pixman=0.40.0=h36c2ea0_0
protobuf=3.12.2=pypi_0
py-opencv=3.4.2=py37hb342d67_1
pycparser=2.20=pyh9f0ad1d_2
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37h05f1152_2
pysoundfile=0.10.3.post1=pyhd3deb0d_0
python=3.7.7=hcff3b4d_5
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.5.1=py3.7_cuda10.2.89_cudnn7.6.5_0
pytz=2020.1=py_0
pyyaml=5.3.1=py37h7b6447c_0
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
scikit-learn=0.23.1=py37h423224d_0
scipy=1.5.0=py37h0b6359f_0
seaborn=0.11.1=hd8ed1ab_1
seaborn-base=0.11.1=pyhd8ed1ab_1
setuptools=47.3.1=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.15.0=py_0
sktime=0.5.3=py37hcd2ae1e_0
sqlite=3.32.3=h62c20be_0
statsmodels=0.12.2=py37h902c9e0_0
tabulate=0.8.7=py37_0
tensorboardx=2.1=pypi_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
torchvision=0.6.1=py37_cu102
tornado=6.0.4=py37h7b6447c_1
tqdm=4.49.0=py_0
tsaug=0.2.1=pypi_0
typing-extensions=3.7.4.3=0
typing_extensions=3.7.4.3=py_0
wheel=0.34.2=py37_0
xorg-fixesproto=5.0=h14c3975_1002
xorg-inputproto=2.3.2=h7f98852_1002
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libx11=1.6.12=h516909a_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxfixes=5.0.3=h516909a_1004
xorg-libxi=1.7.10=h516909a_0
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h7b6447c_0
yaml=0.1.7=had09818_2
zlib=1.2.11=h7b6447c_3
zstd=1.4.4=h0b5b093_3
